app/requirements.txt
openai>=1.23          # official OpenAI SDK
openai-agents>=0.0.16  # Agents SDK - using newer version to fix schema validation issues
google-cloud-firestore>=2.15
google-cloud-pubsub>=2.20
Flask>=3.0
requests>=2.32
beautifulsoup4>=4.12
pydantic>=2.6
python-dotenv>=1.0
